requirements.txt
asyncio
pyzmq
zmq
rcpy
pyubx2
docopt
fastapi
websockets
pyserial==3.5
haversine
pyubx2
numpy
pykalman
pytest
